feat: move media utils into helpers
